Add files via upload
USING WEB TYPOGRAPHY
This HTML and CSS code represents a simple web page for "Payal Fashions," a fashion brand. it comprises web typography .The page includes a title and an image logo at the top. The body features a header with a bold title in blue, styled with the Impact font. Below the header is a navigation bar with links to various sections such as "account," "my cart," and "explore." The main content includes an inspiring quote about fashion and several fashion-related images displayed in a grid format. The footer, styled with an uppercase and cursive font, provides a tagline. The page also features a search button and input box for user interaction.