Add files via upload
practical 3:
AIM: to create a webpage using html and css.
This HTML page provides a list and comparison of the top 4 laptop brands for students. The page includes a title, "Top 5 Laptop Brands for Students," and an ordered list featuring Apple, Dell, Lenovo, and HP laptops. Each brand is highlighted with a brief description of its features. The page also contains a table comparing the specifications of each laptop, including size, RAM type, and price. The total price of all four laptops is displayed at the bottom. Additionally, the page includes a YouTube video about laptops and a video demonstrating laptop features. Styling is applied to the body, including borders, margins, and padding. This page serves as a guide for students looking to purchase a laptop.