Add files via upload
practical 5:
AIM: to create website using web typography.;
This HTML page represents a fashion website called "Payal Fashions." The page includes a stylish layout with a header showcasing the website's logo and a title "Payal Fashions." Below the header, there is a navigation bar with links to account, cart, and contact pages, along with a search bar and a search button. The page features two quotes about fashion, emphasizing its presence in daily life. A series of images showcasing fashion items are displayed in a grid format. The background color is a rich shade of red-brown, complementing the white text. The footer, with an elegant message, is styled in uppercase cursive and appears at the bottom of the page. Overall, the design combines simplicity and elegance to reflect a fashion brand's aesthetic.